requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Django==5.1.1
djangorestframework==3.15.2
drf-extra-fields==3.7.0
dynaconf==3.2.6
filetype==1.2.0
hvac==2.3.0
idna==3.10
knox==0.1.14
loguru==0.7.2
requests==2.32.3
shortuuid==1.0.13
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
validators==0.34.0
win32-setctime==1.1.0
